+++ include/mc_tasks/ExactCubicTrajectoryTask.h
@@ -23,10 +23,6 @@ struct MC_TASKS_DLLAPI ExactCubicTrajectoryTask : public SplineTrajectoryTask<Ex
 {
   friend struct SplineTrajectoryTask<ExactCubicTrajectoryTask>;
 
-public:
-  using SplineTrajectoryBase = SplineTrajectoryTask<ExactCubicTrajectoryTask>;
-
-public:
   /**
    * \brief Trajectory following an exact cubic spline with given initial and
    * final acceleration/velocity. The curve will pass exactly through the