+++ src/mc_control/fsm/states/Message.cpp
@@ -13,34 +13,66 @@ namespace fsm
 
 void MessageState::configure(const mc_rtc::Configuration & config)
 {
+  config("prefix", prefix_);
   config("message", message_);
-  config("type", type_);
+  config("log", logType_);
+  if(config.has("gui"))
+  {
+    gui_ = true;
+    guiCategory_ = config("gui");
+  }
 }
 
-void MessageState::start(Controller & /* ctl */)
+void MessageState::start(Controller & ctl)
 {
-  std::transform(type_.begin(), type_.end(), type_.begin(), [](unsigned char c) { return std::tolower(c); });
+  std::transform(logType_.begin(), logType_.end(), logType_.begin(), [](unsigned char c) { return std::tolower(c); });
 
-  if(type_ == "info")
+  if(logType_.size())
   {
-    LOG_INFO(message_);
+    std::string prefix;
+    std::string message;
+    if(prefix_.size())
+    {
+      prefix = "[" + name() + "::" + prefix_ + "] ";
+    }
+    else
+    {
+      prefix = "[" + name() + "] ";
+    }
+    message = prefix + message_;
+
+    if(logType_ == "info")
+    {
+      LOG_INFO(message);
+    }
+    else if(logType_ == "success")
+    {
+      LOG_SUCCESS(message);
+    }
+    else if(logType_ == "warning")
+    {
+      LOG_WARNING(message);
+    }
+    else if(logType_ == "error")
+    {
+      LOG_ERROR(message);
+    }
+    else if(logType_ == "none")
+    { /* Do not log anything to the terminal */
+    }
+    else
+    {
+      LOG_ERROR(prefix << " Provided log type " << logType_
+                       << " is invalid, assuming info. Supported types are [info, success, warning, error, none]");
+      LOG_INFO(message);
+    }
   }
-  else if(type_ == "success")
+
+  if(gui_)
   {
-    LOG_SUCCESS(message_);
-  }
-  else if(type_ == "warning")
-  {
-    LOG_WARNING(message_);
-  }
-  else if(type_ == "error")
-  {
-    LOG_ERROR(message_);
-  }
-  else
-  {
-    LOG_WARNING("[" << name() << "] Unknown type: " << type_ << ", treating as INFO")
-    LOG_INFO(message_);
+    labelName_ = prefix_.size() ? prefix_ : name();
+    ctl.gui()->addElement(guiCategory_,
+                          mc_rtc::gui::Label(labelName_, [this]() -> const std::string & { return message_; }));
   }
   output("OK");
 }
@@ -48,6 +80,14 @@ void MessageState::start(Controller & /* ctl */)
 bool MessageState::run(Controller &)
 {
   return true;
+}
+
+void MessageState::teardown(Controller & ctl)
+{
+  if(gui_)
+  {
+    ctl.gui()->removeElement(guiCategory_, labelName_);
+  }
 }
 
 } // namespace fsm